requirements_frontend.txt
streamlit>=1.28.2
pandas>=2.1.3
requests>=2.31.0
pillow>=9.0.1
google-cloud-run>=0.10.5